requirements.txt
Pillow>=10.0.0
reportlab>=4.0.0
PyQt6>=6.6.0
